fix: create configcompression
