+++ config/confighttp/confighttp_test.go
@@ -40,6 +40,7 @@ import (
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/configauth"
+	"go.opentelemetry.io/collector/config/configcompression"
 	"go.opentelemetry.io/collector/config/configtls"
 	"go.opentelemetry.io/collector/internal/testutil"
 )
@@ -868,7 +869,7 @@ func TestHTTPClientCompression(t *testing.T) {
 
 	tests := []struct {
 		name        string
-		encoding    CompressionType
+		encoding    configcompression.CompressionType
 		reqBody     []byte
 		shouldError bool
 	}{
@@ -943,7 +944,7 @@ func TestHTTPClientCompression(t *testing.T) {
 			require.NoError(t, err, "failed to create request to test handler")
 
 			client := http.Client{}
-			if tt.encoding != CompressionEmpty && tt.encoding != CompressionNone {
+			if configcompression.IsCompressed(tt.encoding) {
 				client.Transport = newCompressRoundTripper(http.DefaultTransport, tt.encoding)
 			}
 			res, err := client.Do(req)
@@ -1053,67 +1054,6 @@ func TestHTTPContentDecompressionHandler(t *testing.T) {
 				assert.Equal(t, tt.respBody, string(body))
 			}
 			require.NoError(t, srv.Close())
-		})
-	}
-}
-
-func TestUnmarshalText(t *testing.T) {
-	tests := []struct {
-		name            string
-		compressionName []byte
-		shouldError     bool
-	}{
-		{
-			name:            "ValidGzip",
-			compressionName: []byte("gzip"),
-			shouldError:     false,
-		},
-		{
-			name:            "ValidZlib",
-			compressionName: []byte("zlib"),
-			shouldError:     false,
-		},
-		{
-			name:            "ValidDeflate",
-			compressionName: []byte("deflate"),
-			shouldError:     false,
-		},
-		{
-			name:            "ValidSnappy",
-			compressionName: []byte("snappy"),
-			shouldError:     false,
-		},
-		{
-			name:            "ValidZstd",
-			compressionName: []byte("zstd"),
-			shouldError:     false,
-		},
-		{
-			name:            "ValidEmpty",
-			compressionName: []byte(""),
-			shouldError:     false,
-		},
-		{
-			name:            "ValidNone",
-			compressionName: []byte("none"),
-			shouldError:     false,
-		},
-		{
-			name:            "Invalid",
-			compressionName: []byte("ggip"),
-			shouldError:     true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			temp := CompressionNone
-			err := temp.UnmarshalText(tt.compressionName)
-			if tt.shouldError {
-				assert.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			assert.Equal(t, temp, CompressionType(tt.compressionName))
 		})
 	}
 }
